+++ src/components/ElementTiptap.vue
@@ -1,6 +1,8 @@
 <template>
   <div
     v-if="editor"
+    :style="editorSizeStyle"
+    :class="{ 'el-tiptap-editor--fullscreen': isFullscreen }"
     class="el-tiptap-editor"
   >
     <menu-bubble
@@ -40,17 +42,19 @@
 </template>
 
 <script lang="ts">
-import { Component, Prop, Watch, Model, Mixins } from 'vue-property-decorator';
+import { Component, Prop, Watch, Provide, ProvideReactive, Model, Mixins } from 'vue-property-decorator';
 import { Editor, EditorContent, Extension, EditorUpdateEvent } from 'tiptap';
-
 import { Placeholder } from 'tiptap-extensions';
+import { EditorProps } from 'prosemirror-view';
 import ContentAttributes from '@/extensions/content_attributes';
-import { capitalize } from '@/utils/shared';
+import { isNaN, capitalize } from '@/utils/shared';
 import { EVENTS } from '@/constants';
 import i18nMixin from '@/mixins/i18nMixin';
 
 import MenuBar from './MenuBar/index.vue';
 import MenuBubble from './MenuBubble/index.vue';
+
+const DEFAULT_EDITOR_SIZE_UNIT = 'px';
 
 const COMMON_EMIT_EVENTS: EVENTS[] = [
   EVENTS.TRANSACTION,
@@ -87,6 +91,12 @@ export default class ElTiptap extends Mixins(i18nMixin) {
   readonly placeholder!: string;
 
   @Prop({
+    type: Object,
+    default: () => ({}),
+  })
+  readonly editorProps!: EditorProps;
+
+  @Prop({
     type: String,
     default: 'html',
     validator (output): boolean {
@@ -107,6 +117,18 @@ export default class ElTiptap extends Mixins(i18nMixin) {
   })
   readonly spellcheck!: boolean | undefined;
 
+  @Prop({
+    type: [String, Number],
+    default: undefined,
+  })
+  readonly width!: string | number;
+
+  @Prop({
+    type: [String, Number],
+    default: undefined,
+  })
+  readonly height!: string | number;
+
   editor: Editor | null = null;
   emitAfterOnUpdate: boolean = false;
 
@@ -114,6 +136,18 @@ export default class ElTiptap extends Mixins(i18nMixin) {
     if (!this.editor) return 0;
 
     return this.editor.state.doc.textContent.length;
+  }
+
+  get editorSizeStyle () {
+    let { width, height } = this;
+
+    width = isNaN(Number(width)) ? width : `${width}${DEFAULT_EDITOR_SIZE_UNIT}`;
+    height = isNaN(Number(height)) ? height : `${height}${DEFAULT_EDITOR_SIZE_UNIT}`;
+
+    return {
+      width,
+      height,
+    };
   }
 
   @Watch('content')
@@ -147,10 +181,11 @@ export default class ElTiptap extends Mixins(i18nMixin) {
 
     this.editor = new Editor({
       editable: !this.readonly,
+      editorProps: this.editorProps,
       useBuiltInExtensions: false,
       extensions,
-      content: this.content,
       ...eventOptions,
+      content: this.content,
       onUpdate: this.onUpdate.bind(this),
     });
 
@@ -167,7 +202,9 @@ export default class ElTiptap extends Mixins(i18nMixin) {
     let extensions: Extension[] = [];
 
     const spellcheck = this.spellcheck == null
-      ? this.$elementTiptapPlugin.spellcheck
+      ? this.$elementTiptapPlugin
+        ? this.$elementTiptapPlugin.spellcheck
+        : true
       : this.spellcheck;
 
     extensions = extensions.concat(
@@ -207,6 +244,12 @@ export default class ElTiptap extends Mixins(i18nMixin) {
     }
 
     this.$emit(this.genEvent(EVENTS.UPDATE), output, options);
+  }
+
+  // TODO: provide to fullscreen command button, need to be optimized
+  @ProvideReactive() isFullscreen = false;
+  @Provide() toggleFullscreen () {
+    this.isFullscreen = !this.isFullscreen;
   }
 
   private genEvent (event: EVENTS) {