+++ src/components/ExtensionViews/ImageView.vue
@@ -1,10 +1,14 @@
 <template>
-  <span
-    :class="imageViewClass"
-  >
-    <div class="image-view__body">
+  <span :class="imageViewClass">
+    <div
+      :class="{
+        'image-view__body--focused': selected,
+        'image-view__body--resizing': resizing,
+      }"
+      class="image-view__body"
+    >
       <img
-        :src="node.attrs.src"
+        :src="src"
         :title="node.attrs.title"
         :alt="node.attrs.alt"
         :width="width"
@@ -25,27 +29,43 @@
           class="image-resizer__handler"
           @mousedown="onMouseDown($event, direction)"
         />
-
-        <span
-          class="image-view__delete-trigger"
-          @click="removeImage"
-        >
-          <v-icon name="regular/trash-alt" />
-        </span>
       </div>
+
+      <!-- when image is break text or float
+      bubble menu's position is miscalculated
+      use el-popover instead bubble menu -->
+      <el-popover
+        :value="selected"
+        :visible-arrow="false"
+        placement="top"
+        trigger="manual"
+        popper-class="el-tiptap-image-popper"
+      >
+        <image-bubble-menu
+          :node="node"
+          :view="view"
+          :update-attrs="updateAttrs"
+        />
+
+        <div
+          slot="reference"
+          class="image-view__body__placeholder"
+        />
+      </el-popover>
     </div>
   </span>
 </template>
 
 <script lang="ts">
 import { Component, Prop, Vue } from 'vue-property-decorator';
+import { Popover } from 'element-ui';
 import { Node as ProsemirrorNode } from 'prosemirror-model';
 import { NodeSelection } from 'prosemirror-state';
 import { EditorView } from 'prosemirror-view';
-import { deleteSelection } from 'prosemirror-commands';
-import Icon from 'vue-awesome/components/Icon.vue';
-import 'vue-awesome/icons/regular/trash-alt';
-import { resolveImg } from '@/utils/image';
+import { ResizeObserver } from '@juggle/resize-observer';
+import { resolveImg, ImageDisplay } from '@/utils/image';
+import { clamp } from '@/utils/shared';
+import ImageBubbleMenu from '../MenuBubble/ImageBubbleMenu.vue';
 
 const enum ResizeDirection {
   TOP_LEFT = 'tl',
@@ -55,10 +75,12 @@ const enum ResizeDirection {
 };
 
 const MIN_SIZE = 20;
+const MAX_SIZE = 100000;
 
 @Component({
   components: {
-    'v-icon': Icon,
+    [Popover.name]: Popover,
+    ImageBubbleMenu,
   },
 })
 export default class ImageView extends Vue {
@@ -69,7 +91,9 @@ export default class ImageView extends Vue {
   readonly node!: ProsemirrorNode;
 
   @Prop({
-    type: EditorView,
+    // TODO: EditorView type check failed
+    // issue: https://github.com/Leecason/element-tiptap/issues/21#issuecomment-605615966
+    type: Object,
     required: true,
   })
   readonly view!: EditorView;
@@ -92,10 +116,19 @@ export default class ImageView extends Vue {
   })
   readonly selected!: boolean;
 
+  maxSize = {
+    width: MAX_SIZE,
+    height: MAX_SIZE,
+  };
+
   originalSize = {
     width: 0,
     height: 0,
   };
+
+  resizeOb = new ResizeObserver(() => {
+    this.getMaxSize();
+  });
 
   resizeDirections = [
     ResizeDirection.TOP_LEFT,
@@ -126,11 +159,15 @@ export default class ImageView extends Vue {
     return this.node.attrs.height;
   }
 
+  private get display (): ImageDisplay {
+    return this.node.attrs.display;
+  }
+
   private get imageViewClass () {
-    return {
-      'image-view': true,
-      'image-view--focused': this.selected,
-    };
+    return [
+      'image-view',
+      `image-view--${this.display}`,
+    ];
   }
 
   private async created () {
@@ -147,6 +184,14 @@ export default class ImageView extends Vue {
     };
   }
 
+  private mounted () {
+    this.resizeOb.observe(this.view.dom);
+  }
+
+  private beforeDestroy () {
+    this.resizeOb.disconnect();
+  }
+
   // https://github.com/scrumpy/tiptap/issues/361#issuecomment-540299541
   private selectImage () {
     const { state } = this.view;
@@ -156,11 +201,16 @@ export default class ImageView extends Vue {
     this.view.dispatch(tr);
   }
 
-  private removeImage () {
-    const { state, dispatch } = this.view;
-    deleteSelection(state, dispatch);
+  /* invoked when window or editor resize */
+  private getMaxSize () {
+    const { width } = getComputedStyle(this.view.dom);
+    this.maxSize.width = parseInt(width, 10);
   }
 
+  /* on resizer handler mousedown
+   * record the position where the event is triggered and resize direction
+   * calculate the initial width and height of the image
+   */
   private onMouseDown (e: MouseEvent, dir: ResizeDirection): void {
     e.preventDefault();
     e.stopPropagation();
@@ -173,14 +223,19 @@ export default class ImageView extends Vue {
     const aspectRatio = originalWidth / originalHeight;
 
     let { width, height } = this.node.attrs;
+    const maxWidth = this.maxSize.width;
 
     if (width && !height) {
+      width = width > maxWidth ? maxWidth : width;
       height = Math.round(width / aspectRatio);
     } else if (height && !width) {
       width = Math.round(height * aspectRatio);
+      width = width > maxWidth ? maxWidth : width;
     } else if (!width && !height) {
-      width = originalWidth;
-      height = originalHeight;
+      width = originalWidth > maxWidth ? maxWidth : originalWidth;
+      height = Math.round(width / aspectRatio);
+    } else {
+      width = width > maxWidth ? maxWidth : width;
     }
 
     this.resizerState.w = width;
@@ -203,7 +258,7 @@ export default class ImageView extends Vue {
     const dy = (e.clientY - y) * (/t/.test(dir) ? -1 : 1);
 
     this.updateAttrs({
-      width: Math.max(w + dx, MIN_SIZE),
+      width: clamp(w + dx, MIN_SIZE, this.maxSize.width),
       height: Math.max(h + dy, MIN_SIZE),
     });
   }