+++ src/components/MenuCommands/Link/UnlinkCommandButton.vue
@@ -10,7 +10,7 @@
 import { Component, Prop, Mixins } from 'vue-property-decorator';
 import { MenuData } from 'tiptap';
 import i18nMixin from '@/mixins/i18nMixin';
-import CommandButton from './CommandButton.vue';
+import CommandButton from '../CommandButton.vue';
 
 @Component({
   components: {