+++ src/index.ts
@@ -27,18 +27,3 @@ export {
   ElementTiptapPlugin,
   ElementTiptap,
 };
-
-;(function autoInstall () {
-  let globalScope: any = null;
-
-  if (typeof window !== 'undefined') {
-    globalScope = window;
-    // @ts-ignore
-  } else if (typeof global !== 'undefined') {
-    // @ts-ignore
-    globalScope = global;
-  }
-  if (globalScope && globalScope.Vue) {
-    globalScope.Vue.use(ElementTiptapPlugin);
-  }
-})();